services/architecture_suite/requirements.txt
fastapi>=0.95
uvicorn[standard]
SQLAlchemy>=1.4
psycopg2-binary
alembic
pydantic<2.0
python-jose
redis>=4.0
pytest
pytest-cov
flake8
